Add GitHub Actions (#6)
This PR adds configuration for GitHub Actions and updates the VS Code and devcontainer configs to match the tasks available in GitHub Actions. In particular, this adds several linters and `pytest-cov` as development dependencies. The PR also updates the gevent dependency as all non-current versions of gevent are broken in python 3.10.