Join greenlet in parallel thread (#10)
Fixes #8. This PR changes the behaviour of `asyncio.greenlet_to_future` by making sure that the greenlet is awaited in a separate thread which prevents blocking of the main thread and makes interrupts safe to use with the code. The PR also adds a note to the README describing how to work with `gevent.sleep`.