Fix missing files in build (#25)
This PR fixes the build tool config in pyproject.toml to include the asyncio_gevent package directory.

PDM would normally auto-detect and auto-include packages, but because I set the `includes` field to include the `NOTICE` file, the auto-detecting was turned off, so the package directory was no longer included.

The PR also adds a check that runs during CI runs on PR, push to main and before every release to ensure the correct files are included in the build as well as associated tests for the script.

Fixes #24.